Fix lurking truncation bugs in ch1 exercises

+++ ch01-intro/examples/FahrenheitToCelsius/main.c
@@ -5,12 +5,12 @@ int main() {
   int lower = 0;
   int upper = 300;
   int step = 20;
-  float fahr = lower;
+  int fahr = lower;
   float celsius;
 
   while (fahr <= upper) {
-    celsius = (5.0 / 9.0) * (fahr - 32.0);
-    printf("%11.0f | %8.1f\n", fahr, celsius);
+    celsius = (5.0f / 9.0f) * (fahr - 32.0f);
+    printf("%11i | %8.1f\n", fahr, celsius);
     fahr += step;
   }
